Major rehaul:  Allow selection between output to file and to sox
 + Add a high level function 'add_named_note' to the voice
 + Add a 'write' method to the muse which prints all samples
   of a voice

TODO: For continuous operation, we need to print to the player's
handle.  We can _not_ do this for output to a .wav file because this
needs the length in the header!
We can, however, use the SoX output capabilities, then it is the
problem of SoX to keep the stuff in memory.