Feature: The format specification allows to omit repeated note durations
This is a feature stolen from Lilypond:  For notes with the same
duration or octave, only the first duration needs be written
This considerably shortens many lines.

In contrast to Lilypond we use numbers for octaves, and actual
durations and not their inverse: Lilypond writes quarter notes
as "4", we write them as "1/4".