WIP: Eliminate rate() from Units
There's work to be done for Voice and the mrt_play stuff which needs to
know about rate and bpm (or, precisely, samples per full note).  This can
change during the lifetime of a voice.

Also, notes_per_second was a misnomer and has been reversed to
seconds_per_note.