WIP/Feature: Add a "Register" widget with a bunch of stops
Still WIP because we want a UI to create a sine wave with overtones,
and to store its parameters.  We want to play it with a given frequency.

Later we want to add a "key" to play it on keypress.  Only then we can
have an envelope.

A few effects (tremolo, vibrato) could be added before we have a key.