Bugfix / Cleanup: Silence and update triggers
"$silent" is now part of the player only.
Key press and release now trigger an immediate update, not waiting for the
next timer event.  The tick does no longer update the server's next_sample,
this is done every time $server->current_sample is called instead.

There is a noticeable delay between a key press and the corresponding sound.
Perhaps the GUI has too many events?

Nope.  Per experiment it is not faster with only 10 overtones.