WIP interactive envelope: Relevant progress
Attack, decay and sustain seem to work as intended.
Release is not yet implemented.

The first two "power" and "play" buttons are gone, the program
is active from its start.

There are rough edges between "endless loop" and "play key" and
a few more, but this state is supposed to be saved.