WIP Bugfix and still fighting with release
Turned out that mrt_play and the AA::Voice had a nasty bug which
failed to synchronize tracks sarting or ending within a music roll.

Also, release still isn't done correctly for mrt_play.  A carry
should propagate from one MRT section to the next.  I need to check
where this goes wrong.

Another (minor) bug which is now "almost" fixed is how an output
file name is passed through the objects.  Should it be an object
property or rather an argument for method calls?  _Some_
stuff needs the output handle to be the same for all method
cals.