Yet another WIP: AA::Voice is an AA::Source
AA::Voice does not yet use AA::Server but it could serve as a
source for interactive replay - once the samples have been prepared.
Preparation happens from addong notes - or samples.

The "close_function", so far unused, has been killed.  We shall
try to add the envelope as one effect in a list.

I wonder what this will do to the processing of "carry" for notes.
Probably this will be retained because it happens in add_notes