Feature: bin/spectre - change config values via UI
This propably needs some safeguards.  Changing the values
should not be done while sound is playing.  Also, it is easily
possible to get your CPU load to 100% with too narrow resolution
and/or too high frame rate.  That depends on the hardware,
of course.

* bin/spectre: Allow the frame rate, spectral resolution and max
frequency (for display) to be changed via GUI.

* lib/Audio/Aoede/Player/SoX.pm (close_pipe): Safeguard against
calling close_pipe twice.

* lib/Audio/Aoede/UI/Spectrum.pm (set_limit): New method to change
the display limit.

* lib/Audio/Aoede/UI/Spectre/Options.pm: New module for changing
configuration values.