Features: Add "record the microphone" and "save as"
Also, some rearrangements and cleanup.

* bin/spectre: Add configuration variables as command line parameters
Add support for a recorder based on Audio::Portaudio.  A player is
there, too, but we don't use it.
Refactoring: Push the timer and user options into their own modules.
Add "Save as" dialog, using a SoX player for support of various audio
formats.
Add a button to start recording from the microphone (which uses the
portaudio recorder)

* lib/Audio/Aoede/Recorder/PortAudio.pm: A new recorder based on
Audio::PortAudio.  This is a "low level" library, no extra
process.  We are silencing a bunch of errors which are harmless /
to be expected (ALSA config) or just wrong messages.

* lib/Audio/Aoede/Player/Ecasound.pm: Yet another player.  Unused.
The timer based pipe creates underruns with Ecasound.  Might work
as a "save as"-player, untested.

* lib/Audio/Aoede/Player/PortAudio.pm: A player based on
Audio::PortAudio.  Unused right now.  Does not support "save as".

* lib/Audio/Aoede/UI/Spectre/Options.pm: Handle the UI to change
spectrum refresh rate, resolution, and frequency limit.

* lib/Audio/Aoede/UI/Spectre/Timer.pm: Factor out the timer.
During the tests this module lost some of its fat, it is rather
slim now.