Bookkeeping: Switch over to Makefile.PL build
The SoundFont synthesizer now wants to support the generator
modEnvToPitch which I implemented as a PDL::PP helper to get
decent performance.  In contrast to PDL::Filter::Biquad this
helper routine is very specific to our project and should not
occupy the PDL namespace.

Also, the pure-perl biquad filter, so far sitting in
lib/Audio/Aoede/Filter/Biquad.pm, is about to be deleted.